+++ src/views/settings/GraphSettingsView.ts
@@ -5,6 +5,8 @@ import { GroupSettings } from "../../settings/categories/GroupSettings";
 import { DisplaySettings } from "../../settings/categories/DisplaySettings";
 import { ClusterPhysicsSettings } from "../../settings/categories/ClusterPhysicsSettings";
 import ClusterPhysicsSettingsView from "./categories/ClusterPhysicsSettingsView";
+import { McpSettings } from "../../settings/categories/McpSettings";
+import McpSettingsView from "./categories/McpSettingsView";
 import { ExtraButtonComponent } from "obsidian";
 import State, { StateChange } from "../../util/State";
 import EventBus from "../../util/EventBus";
@@ -76,6 +78,11 @@ export class GraphSettingsView extends HTMLDivElement {
 			this.settingsState.createSubState("value.clusterPhysics", ClusterPhysicsSettings),
 			"Cluster Physics",
 			ClusterPhysicsSettingsView
+		);
+		this.appendSetting(
+			this.settingsState.createSubState("value.mcp", McpSettings),
+			"MCP Server",
+			McpSettingsView
 		);
 		this.initListeners();
 		this.toggleCollapsed(this.isCollapsedState.value);